pyproject.toml
[project]
name = "agenticlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-tavily>=0.2.3",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.6",
    "langsmith>=0.3.45",
    "python-dotenv>=1.1.0",
]
